requirements.txt
flask
flask_cors
numpy
tensorflow
pillow
google-generativeai
